cleanup

+++ Makefile
@@ -1,34 +1,36 @@
-image_name := pg_es_fdw
+compose_project_name := pges
+image_postgres := $(compose_project_name)_postgres
+image_elasticsearch := $(compose_project_name)_elasticsearch
 # docker compose convention
-container_postgres := postgresesfdw_postgres_1
-container_elasticsearch := postgresesfdw_elasticsearch_1
+container_postgres := $(compose_project_name)_postgres_1
+container_elasticsearch := $(compose_project_name)_elasticsearch_1
 # if the containers run, those contain their IP addresses
-ip_postgres := $(shell docker inspect --format '{{ .NetworkSettings.Networks.postgresesfdw_default.IPAddress }}' $(container_postgres))
-ip_elasticsearch := $(shell docker inspect --format '{{ .NetworkSettings.Networks.postgresesfdw_default.IPAddress }}' $(container_elasticsearch))
+ip_postgres := $(shell docker inspect --format '{{ .NetworkSettings.Networks.$(compose_project_name)_default.IPAddress }}' $(container_postgres) 2>/dev/null)
+ip_elasticsearch := $(shell docker inspect --format '{{ .NetworkSettings.Networks.$(compose_project_name)_default.IPAddress }}' $(container_elasticsearch) 2>/dev/null)
 
-default: build
+# Command Shortcuts
+# =================
+compose := docker-compose --project-name $(compose_project_name)
+# password is set in docker-compose.yml
+psql := PGPASSWORD=foobarbaz psql -h $(ip_postgres) -U postgres
+
+# default target
+# ==============
+default: compose
 
 # used in ADD clause in Dockerfile
 esfdw:
 	git submodule update --init
 
+compose:
+	$(compose) up
+
 build: esfdw
-	docker build -t $(image_name) .
+	$(compose) build
 
-run:
-	docker run --rm -e POSTGRES_PASSWORD=geheim --name $(container_postgres) $(image_name) 
-
-# if started via run, this is how you get the IP address:
-#psql -h $(shell docker inspect --format '{{ .NetworkSettings.IPAddress }}' $(container_postgres)) -U postgres
-# docker compose creates a network called postgresesfdw_default
 connect:
-	psql -h $(ip_postgres) -U postgres
+	$(psql)
 
-# psql config file with password for postgres. See POSTGRES_PASSWORD in run.
-.PHONY: pgpass
-pgpass:
-	cat pgpass >> ~/.pgpass
-	chmod 600 ~/.pgpass
 # our container needs to have python 2 for esfdw to work
 check_python_version:
 	docker run -it --rm $(image_name) python -V | grep 'Python 2'
@@ -36,9 +38,10 @@ check_python_version:
 debug:
 	docker run -it --rm $(image_name) bash
 
-# elasticsearch example data
+# Elasticsearch Example Data
+# ==========================
 accounts.zip:
-	wget 'https://github.com/bly2k/files/blob/master/accounts.zip?raw=true' -O accounts.zip
+	test -f accounts.zip || wget 'https://github.com/bly2k/files/blob/master/accounts.zip?raw=true' -O accounts.zip
 accounts.json: accounts.zip
 	test -f accounts.json || unzip accounts.zip
 load_es_example_data: accounts.json
@@ -48,16 +51,14 @@ show_es_indices:
 
 # use FDW
 # =======
-psql := psql -h $(ip_postgres) -U postgres
 create_extension:
 	$(psql) -c "CREATE EXTENSION IF NOT EXISTS multicorn;"
 create_server_es:
 	$(psql) < create_server.sql
-
 get_mapping_for_our_data:
 	docker exec $(container_elasticsearch) curl localhost:9200/bank/account/_mapping > mapping.json
 generate_foreign_table_create_statement:
-	cat mapping.json | docker exec -i postgresesfdw_postgres_1 python -m esfdw.mapping_to_schema -i bank -d account -s es > foreign_table.sql
+	cat mapping.json | docker exec -i $(compose_project_name)_postgres_1 python -m esfdw.mapping_to_schema -i bank -d account -s es > foreign_table.sql
 	# remove the last option line (column_name_translation)
 	perl -pe "s/\s+column_name_translation 'true'\n//; s/ index 'bank',/ index 'bank'/" -i foreign_table.sql
 create_foreign_table:
